GetAllQuery and FindAllQuery to return IQueryable
These are internal methods and by returning IQueryable we can keep it as
a query (and on the server side if the LINQ provider does it that way)
until the last minute.  Previously we were returning IEnumerable and had
a call to ToList() in these methods and so when doing a selector, each
entire entity was returned and then the selector was supplied.  This
wasn't ideal because more data was transferred over the wire than needed
to be.

So now the FindAll and QueryAll that take a selector, call the
GetAllQuery or FindAllQuery which keeps it as a Quyerable and then adds
the .Select() part so it's part of the query.  Then the FindAll or
GetAll calls ToList() to take it from the server side and returne the
actual results.